[changelog] Core: Added support for sub networks, using CompositeProcessor and Sink/Source processors

+++ src/qt/editor/processorlistwidget.cpp
@@ -392,7 +392,7 @@ void ProcessorTreeWidget::extractInfoAndAddProcessor(ProcessorFactoryObject* pro
         case 5: {  // Most Used
             auto it = useCounts_.find(processor->getClassIdentifier());
             if (it != useCounts_.end()) {
-                sortVal.prepend(QVariant::fromValue<qint64>(-it->second));
+                sortVal.prepend(QVariant::fromValue<qint64>(-static_cast<qint64>(it->second)));
             } else {
                 sortVal.prepend(0);
             }
